Primary Keys are now reset upon each launch

+++ app/src/main/java/com/scottb4/lunchilicious/data/MenuItemEntityClass.kt
@@ -5,8 +5,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "menu_item")
 data class MenuItem (
-    @PrimaryKey(autoGenerate = true)
-    val id: Int = 0,
+    @PrimaryKey(autoGenerate = true) val id: Long = 0,
     val type: String,
     val name: String,
     val description: String,
